requirements.txt
boto3==1.10.32
botocore==1.13.32
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-rest-knox==4.1.0
django-storages==1.8
djangorestframework==3.10.3
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
jmespath==0.9.4
Pillow==6.2.0
pipfile==0.0.2
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-decouple==3.3
pytz==2019.2
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.5
waitress==1.3.1
whitenoise==4.1.3
